refactor: split bot.js into modules
