+++ bot/state.js
@@ -1,0 +1,4 @@
+const userState = {};
+const domainState = {};
+
+module.exports = { userState, domainState };
